Move update tracked vals back
