+++ beacon-chain/monitor/process_block.go
@@ -56,26 +56,6 @@ func (s *Service) processBlock(ctx context.Context, b block.SignedBeaconBlock) {
 	s.processAttestations(ctx, state, blk)
 }
 
-// updateSyncCommitteeTrackedVals updates the sync committee assignments of our
-// tracked validators. It gets called when we sync a block after the Sync Period changes.
-func (s *Service) updateSyncCommitteeTrackedVals(state state.BeaconState) {
-	s.Lock()
-	defer s.Unlock()
-	for idx := range s.trackedValidators {
-		syncIdx, err := helpers.CurrentPeriodSyncSubcommitteeIndices(state, idx)
-		if err != nil {
-			log.WithError(err).WithField("ValidatorIndex", idx).Error(
-				"Sync committee assignments will not be reported")
-			delete(s.trackedSyncCommitteeIndices, idx)
-		} else if len(syncIdx) == 0 {
-			delete(s.trackedSyncCommitteeIndices, idx)
-		} else {
-			s.trackedSyncCommitteeIndices[idx] = syncIdx
-		}
-	}
-	s.lastSyncedEpoch = slots.ToEpoch(state.Slot())
-}
-
 // processProposedBlock logs the event that one of our tracked validators proposed a block that was included
 func (s *Service) processProposedBlock(state state.BeaconState, root [32]byte, blk block.BeaconBlock) {
